requirements.txt
pillow==10.1.0
pyautogui==0.9.54
pydirectinput==1.0.4
pynput==1.7.6
pyscreeze==0.1.29
opencv-python==4.8.0.74
